Move project carousel into component module
